Resolución Ejercicio9 [Efimov Valentina]

+++ src/ar/edu/unju/escmi/tp2/ejercicio9/Ejercicio9.java
@@ -6,12 +6,13 @@ public class Ejercicio9 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Ingrese un año: ");
-        int anio = sc.nextInt();
+        short anio = sc.nextShort();
+
         if ((anio%4 == 0 && anio%100 != 0) || anio%400 == 0)
          System.out.println("El año " +anio+ " es bisiesto");
         else
          System.out.println("El año " +anio+ " no es bisiesto");
-        
+    
         sc.close();
     }
 }
